requirements.txt
pydexcom==0.4.1
pystray>=0.19.0
Pillow>=9.0.0